split key commands and text expansions into separate files for easier file searches
