+++ lib/gh-koenig/addon/components/gh-koenig.js
@@ -5,7 +5,8 @@ import layout from '../templates/components/gh-koenig';
 import Mobiledoc from 'mobiledoc-kit';
 import {MOBILEDOC_VERSION} from 'mobiledoc-kit/renderers/mobiledoc';
 import createCardFactory from '../lib/card-factory';
-import defaultCommands from '../options/default-commands';
+import registerKeyCommands from '../options/key-commands';
+import registerTextExpansions from '../options/text-expansions';
 import editorCards  from '../cards/index';
 import {getCardFromDoc, checkIfClickEventShouldCloseCard, getPositionOnScreenFromRange} from '../lib/utils';
 import counter from 'ghost-admin/utils/word-count';
@@ -137,7 +138,10 @@ export default Component.extend({
         // set global editor for debugging and testing.
         window.editor = editor;
 
-        defaultCommands(editor); // initialise the custom text handlers for MD, etc.
+        // set up our default key handling and text expansions to emulate MD behaviour
+        registerKeyCommands(editor);
+        registerTextExpansions(editor);
+
         // shouldFocusEditor is only true when transitioning from new to edit, otherwise it's false or undefined.
         // therefore, if it's true it's after the first lot of content is entered and we expect the caret to be at the
         // end of the document.
